+++ lib/internal/http2/core.js
@@ -671,8 +671,6 @@ function submitShutdown(options) {
 
 function finishSessionDestroy(self, socket) {
   const state = self[kState];
-  if (state.destroyed)
-    return;
 
   if (!socket.destroyed)
     socket.destroy();
@@ -955,6 +953,7 @@ class Http2Session extends EventEmitter {
       return;
     debug(`[${sessionName(this[kType])}] destroying nghttp2session`);
     state.destroying = true;
+    state.destroyed = false;
 
     // Unenroll the timer
     this.setTimeout(0, sessionOnTimeout);
@@ -968,9 +967,6 @@ class Http2Session extends EventEmitter {
     // socket.pause();
     delete this[kSocket];
     delete this[kServer];
-
-    state.destroyed = false;
-    state.destroying = true;
 
     if (this[kHandle] !== undefined)
       this[kHandle].destroying();
